+++ .gitignore
@@ -1,4 +1,3 @@
-test.c
-p
-vdir
 docs/TASKS.md
+artefacts
+snapdir
