+++ src/main.h
@@ -1,0 +1,10 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+#define MAX_PATH 1024
+#define MAX_STAT 1024
+#define BUFFER 4096
+#define MIN_INPUT_DIRS 5
+#define MAX_INPUT_DIRS 14
+
+#endif
